Extract versioning code to semver package
