+++ internal/registry/quay/client.go
@@ -11,6 +11,7 @@ import (
 	"github.com/AlexGustafsson/cupdate/internal/httputil"
 	"github.com/AlexGustafsson/cupdate/internal/registry"
 	"github.com/AlexGustafsson/cupdate/internal/registry/oci"
+	"github.com/AlexGustafsson/cupdate/internal/semver"
 )
 
 type Client struct {
@@ -27,7 +28,7 @@ func (c *Client) GetLatestVersion(ctx context.Context, image oci.Reference) (*re
 		return nil, nil
 	}
 
-	currentVersion, err := oci.ParseVersion(image.Tag)
+	currentVersion, err := semver.ParseVersion(image.Tag)
 	if err != nil {
 		return nil, fmt.Errorf("unsupported version: %w", err)
 	} else if currentVersion == nil {
@@ -77,7 +78,7 @@ func (c *Client) GetLatestVersion(ctx context.Context, image oci.Reference) (*re
 			continue
 		}
 
-		newVersion, err := oci.ParseVersion(tag.Name)
+		newVersion, err := semver.ParseVersion(tag.Name)
 		if err != nil || newVersion == nil {
 			continue
 		}
