+++ internal/semver/semver_test.go
@@ -1,4 +1,4 @@
-package oci
+package semver
 
 import (
 	"fmt"
